requirements.txt
fastapi[standard]==0.136.0
anthropic==0.96.0
streamlit==1.45.0
python-dotenv==1.1.0